+++ maro/rl_v31/training/algorithms/sac.py
@@ -53,8 +53,8 @@ class SACOps(BaseTrainOps):
             target_q = torch.min(target_q1, target_q2)
             y = (reward + self._reward_discount * (1.0 - terminal) * (target_q - self._entropy_coef * next_logps)).float()
 
-        q1 = self._q_critic1(obs, act).float()
-        q2 = self._q_critic2(obs, act).float()
+        q1 = self._q_critic1(obs, act)
+        q2 = self._q_critic2(obs, act)
         loss_q1 = self._q_value_loss_func(q1, y)
         loss_q2 = self._q_value_loss_func(q2, y)
         return loss_q1, loss_q2
